Moved p1_print/p1_capture print functions to new print_utils file.
